requirements.txt
flask
flask_restful
flask-cors
python-dotenv
redis